feat: user system hardening + marketplace Phase 5 + mdgateway fixes
- Auth: MT identity login, password reset flow, token version invalidation
- Marketplace: bundles, fee tiers, decay detector, strategy optimizer, coupons, trials
- Marketplace: refund requests, featured strategies, version history
- MDGateway: broker rediscovery, MT4/MT5 connection extras, SSE limiter
- Frontend: ResetPassword page, BundleTab, FeeTierPanel, OptimizationTab, VersionHistoryTab
- i18n: updated all 5 locales (en/ja/vi/zh-cn/zh-tw)
- Fix: removed dead tests (stripSuffix, cgroupMemoryLimit, currentMemoryRatio)
- Migrations 228-235 (optimization tasks, bundles, fee tiers, decay, settlements, indexes)

Verified: go build ✅ | go vet ✅ | go test ✅

+++ backend/cmd/coldsign/main.go
@@ -159,7 +159,7 @@ func signTx(tx *antv1.UnsignedTx, seed []byte, coldWalletAddr string, credDB *Cr
 	var err error
 	switch tx.Tx.(type) {
 	case *antv1.UnsignedTx_Transfer:
-		sk, err = hdwallet.DeriveDepositPrivKey(seed, uint32(tx.DerivationIndex))
+		sk, err = hdwallet.DeriveDepositPrivKey(seed, tx.DerivationIndex)
 		if err != nil {
 			return nil, fmt.Errorf("derive private key at index %d: %w", tx.DerivationIndex, err)
 		}
